[MOVE/#1] entity 패키지로 User 이동
