+++ app/src/main/res/layout/activity_login.xml
@@ -74,7 +74,8 @@
             android:inputType="textPassword"
             android:paddingVertical="12dp"
             android:text="@={vm.pwd}"
-            app:layout_constraintTop_toBottomOf="@id/tv_login_pwd_label" />
+            app:layout_constraintTop_toBottomOf="@id/tv_login_pwd_label"
+            tools:layout_editor_absoluteX="40dp" />
 
         <Button
             android:id="@+id/btn_login_login"
